# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Api for Autogen.",
        "title": "Autogen API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/admin/auth": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "List User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search Parameter",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting Parameter",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_ListUserResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_ListUserResp"
                        }
                    }
                }
            }
        },
        "/admin/auth/:user_id": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_DeleteUserResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_DeleteUserResp"
                        }
                    }
                }
            }
        },
        "/admin/auth/:user_id/activation": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User Activation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Activation",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_UserActivationResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_UserActivationResp"
                        }
                    }
                }
            }
        },
        "/admin/auth/:user_id/deactivation": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User DeActivation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User DeActivation",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_UserDeActivationResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_UserDeActivationResp"
                        }
                    }
                }
            }
        },
        "/public/auth/login": {
            "post": {
                "description": "Auth Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Auth Login",
                "parameters": [
                    {
                        "description": "Auth Request Parameter",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logisfy_internal_model_request.AuthUserReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_AuthUserResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_AuthUserResp"
                        }
                    }
                }
            }
        },
        "/public/auth/register": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create User (Register) for User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Create User (Register)",
                "parameters": [
                    {
                        "description": "Register User Request Parameter",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logisfy_internal_model_request.RegisterUserReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_RegisterUserResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_RegisterUserResp"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Find an Account by Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Find",
                "responses": {
                    "201": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_AuthUserResp"
                        }
                    },
                    "400": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/logisfy_core_response.ApiResponse-logisfy_internal_model_response_AuthUserResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_AuthUserResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.AuthUserResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
...
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_DeleteUserResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.DeleteUserResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
...
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_ListUserResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.ListUserResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_RegisterUserResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.RegisterUserResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
...
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_UserActivationResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.UserActivationResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
        "logisfy_core_response.ApiResponse-logisfy_internal_model_response_UserDeActivationResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logisfy_internal_model_response.UserDeActivationResp"
                },
                "error": {
                    "type": "string"
                },
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "tin": {
                    "type": "string"
                },
                "tout": {
                    "type": "string"
                }
            }
        },
        "logisfy_internal_model_request.AuthUserReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "remember_me": {
                    "type": "boolean"
                }
            }
        },
...
        "logisfy_internal_model_request.RegisterUserReq": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "UP3": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "unit_induk": {
                    "type": "string"
                }
            }
        },
        "logisfy_internal_model_response.AuthUserResp": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "logisfy_internal_model_response.DeleteUserResp": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "logisfy_internal_model_response.ListUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unit_induk": {
                    "type": "string"
                },
                "up3": {
                    "type": "string"
                }
            }
        },
        "logisfy_internal_model_response.ListUserResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logisfy_internal_model_response.ListUser"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "logisfy_internal_model_response.RegisterUserResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "logisfy_internal_model_response.UserActivationResp": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "logisfy_internal_model_response.UserDeActivationResp": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}